requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
antlr4-python3-runtime==4.8
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopage==0.5.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
cityscapesScripts==2.2.0
click==8.0.3
cliff==3.10.0
cmaes==0.8.2
cmd2==2.3.3
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.6.0
commonmark==0.9.1
configparser==5.2.0
coremltools==5.1.0
-e git+https://github.com/apple/ml-cvnets.git@d38a116fe134a8cd5db18670764fdaafd39a5d4f#egg=cvnets
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
docker-pycreds==0.4.0
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
flake8==4.0.1
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.6.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.43.0
humanfriendly==10.0
hydra-colorlog==1.1.0
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.0
identify==2.4.7
idna==3.3
imageio==2.14.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nodeenv==1.6.0
notebook==6.4.8
numpy==1.19.5
oauthlib==3.2.0
omegaconf==2.1.1
opencv-contrib-python==4.5.5.62
optuna==2.4.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
prettytable==3.0.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pudb==2022.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pyDeprecate==0.3.1
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
pyperclip==1.8.2
pyquaternion==0.9.9
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytorch-lightning==1.5.9
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==11.1.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.4
sh==1.14.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.31
stack-data==0.1.4
stevedore==3.5.0
subprocess32==3.5.4
sympy==1.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.5.4
toml==0.10.2
tomli==2.0.0
torch==1.10.2+cu113
torchmetrics==0.7.0
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing==3.7.4.3
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.8
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.13.0
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
